WhiteFactory and BlackFactory

+++ factory/BlackFactory.java
@@ -7,7 +7,7 @@ public class BlackFactory implements AbstractFactory {
     public BlackTriangle createTriangle(double x1,double y1,double x2,double y2,double x3,double y3) {
         return new BlackTriangle(x1,y1,x2,y2,x3,y3);
     }
-    public BlackCircle createCircle(float center_x,float center_y,float radius) {
+    public BlackCircle createCircle(double center_x,double center_y,double radius) {
         return new BlackCircle(center_x,center_y,radius);
     }
 }
